cleanup and progress editor post
